python/app/requirements.txt
flask
qrcode
pillow
pandas
openpyxl
tqdm
